Crate webb_relayer_handler_utils

Structs

IpInformationResponse

Representation for IP address response

WebbI128

A wrapper type around i128 that implements a correct way for Serialize and Deserialize.

WebbI256

A wrapper type around [I256] that implements a correct way for Serialize and Deserialize.

Enums

Command

Type of Command to use

CommandResponse

Enumerates the command responses

EvmCommandType

Enumerates the supported evm commands for relaying transactions

NetworkStatus

Enumerates the network status response of the relayer

SubstrateCommandType

Enumerates the supported substrate commands for relaying transactions

WithdrawStatus

Enumerates the withdraw status response of the relayer

Functions

into_withdraw_error

A helper function to extract the error code and the reason from EVM errors.

Type Definitions

CommandStream

Type alias for mpsc::Sender

EvmVanchorCommand

The command type for EVM vanchor transactions

SubstrateVAchorCommand

The command type for Substrate vanchor txes